Refactor EmptyChart into its own file/component
