+++ src/components/EmptyChart/EmptyChart.module.scss
@@ -1,0 +1,13 @@
+.emptyChart {
+
+  display:grid;
+  place-items: center;
+  background-color: #eee;
+  color: #666;
+  width: 100%;
+  height: 100%;
+
+  .title {
+    align-self: flex-start;
+  }
+}
