+++ src/components/RepositoryDetail/RepositoryDetail.module.scss
@@ -90,16 +90,3 @@
 
   @include header-style();
 }
-.emptyChart {
-
-  display:grid;
-  place-items: center;
-  background-color: #eee;
-  color: #666;
-  width: 100%;
-  height: 100%;
-
-  .title {
-    align-self: flex-start;
-  }
-}
